% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_methods.R
\name{coef.jm}
\alias{coef.jm}
\title{Fixed Effects Estimates for Survival Submodel in Joint Models}
\usage{
\method{coef}{jm}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{"jm"}.}

\item{...}{additional arguments; currently none is used.}
}
\value{
A list with the elements: 
\tabular{ll}{
\code{gammas} \tab estimated baseline fixed effects. \cr
\code{association} \tab estimated association parameters.   
}
}
\description{
Extracts estimated fixed effects for the event process from a fitted joint model.
}
\examples{
\dontrun{
# linear mixed model fit
fit_lme <- lme(sqrt(CD4) ~ obstime * drug, random = ~ 1 + obstime | patient, 
               data = aids)

# cox model fit
fit_cox <- coxph(Surv(Time, death) ~ drug, data = aids.id)

# joint model fit
fit_jm <- jm(fit_cox, fit_lme, time_var = "obstime")

# fixed effects for the event process 
coef(fit_jm)
}
}
\seealso{
\code{\link{fixef}}, \code{\link{ranef}}, \code{\link{jm}}.
}
\author{
Dimitris Rizopoulos, \email{d.rizopoulos@erasmusmc.nl}.
}
