% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_methods.R
\name{compare_jm}
\alias{compare_jm}
\title{Criteria to compare Joint Models}
\usage{
compare_jm(
  ...,
  type = c("marginal", "conditional"),
  order = c("WAIC", "DIC", "LPML", "none")
)
}
\arguments{
\item{...}{two or more objects inheriting from class \code{"jm"}.}

\item{type}{which log-likelihood function use to calculate the criteria 
(marginal or conditional).}

\item{order}{which criteria use to sort the models in the output.}
}
\value{
An object of class \code{compare_jm}. This is a list with the elements: 
\tabular{ll}{
\code{table} \tab a table with the criteria calculted to each joint model. \cr
\code{type} \tab the log-likelihood function used to calculate the criteria.   
}
}
\description{
Compares two or more fitted joint models using the criteria WAIC, DIC, and LPML.
}
\examples{
\dontrun{
# linear mixed model fits
fit_lme1 <- lme(sqrt(CD4) ~ obstime, random = ~ 1 + obstime | patient, 
                data = aids)

fit_lme2 <- lme(sqrt(CD4) ~ obstime + drug, random = ~ 1 + obstime | patient, 
                data = aids)

# cox model fit
fit_cox <- coxph(Surv(Time, death) ~ drug, data = aids.id)

# joint model fit 1
fit_jm <- jm(fit_cox, fit_lme1, time_var = "obstime")

# joint model fit 2
fit_jm2 <- jm(fit_cox, fit_lme2, time_var = "obstime")

# compare the two fitted joint models
compare_jm(fit_jm1, fit_jm2)
}
}
\seealso{
\code{\link{jm}}.
}
\author{
Dimitris Rizopoulos, \email{d.rizopoulos@erasmusmc.nl}.
}
