% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_methods.R
\name{cumuplot.jm}
\alias{cumuplot.jm}
\title{Cumulative Quantile Plot for Joint Models}
\usage{
\method{cumuplot}{jm}(
  object,
  parm = c("all", "betas", "sigmas", "D", "bs_gammas", "tau_bs_gammas", "gammas",
    "alphas"),
  ...
)
}
\arguments{
\item{object}{an object inheriting from class \code{"jm"}.}

\item{parm}{either \code{"all"} or one specific joint model parameter of interest.}

\item{...}{further arguments passed to \code{\link[coda]{cumuplot}}.}
}
\description{
Plots the evolution of the sample quantiles vs. iterations in a fitted joint 
model.
}
\examples{
\dontrun{
# linear mixed model fit
fit_lme <- lme(sqrt(CD4) ~ obstime * drug, random = ~ 1 + obstime | patient, 
               data = aids)

# cox model fit
fit_cox <- coxph(Surv(Time, death) ~ drug, data = aids.id)

# joint model fit
fit_jm <- jm(fit_cox, fit_lme, time_var = "obstime")

# cumulative quantile plot for the fixed effects in all linear mixed submodels 
cumuplot(fit_jm, parm = "betas")
}
}
\seealso{
\code{\link[coda]{cumuplot}}, \code{\link{jm}}.
}
\author{
Dimitris Rizopoulos, \email{d.rizopoulos@erasmusmc.nl}.
}
