% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_methods.R
\name{family.jm}
\alias{family.jm}
\title{Family Objects for Joint Models}
\usage{
\method{family}{jm}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{"jm"}.}

\item{...}{additional arguments; currently none is used.}
}
\value{
A list of \code{family} objects.
}
\description{
Extracts the error distribution and link function used in the linear
mixed submodel(s) from a fitted joint model.
}
\examples{
\dontrun{
# linear mixed model fits
fit_lme <- lme(log(serBilir) ~ year * sex, random = ~ year | id, 
               data = pbc2)

fit_mm <- mixed_model(ascites ~ year * sex, random = ~ 1 | id, 
                      family = binomial(), data = pbc2)

# cox model fit
fit_cox <- coxph(Surv(years, status2) ~ age, data = pbc2.id)

# joint model fit
fit_jm <- jm(fit_cox, list(fit_lme, fit_mm), time_var = "year")

# family objects from all linear mixed submodels
family(fit_jm)
}
}
\seealso{
\code{\link[stats]{family}}, \code{\link{jm}}.
}
\author{
Dimitris Rizopoulos, \email{d.rizopoulos@erasmusmc.nl}.
}
