% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_methods.R
\name{fixef.jm}
\alias{fixef.jm}
\title{Fixed Effects Estimates for Linear Mixed Submodels in Joint Models}
\usage{
\method{fixef}{jm}(object, outcome = 1, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{"jm"}.}

\item{outcome}{the linear mixed submodel to extract the estimated fixed effects. 
If greater than the total number of linear mixed submodels, extracts from all 
of them.}

\item{...}{additional arguments; currently none is used.}
}
\value{
A numeric vector of the estimated fixed effects for the 
\code{outcome} selected. If \code{outcome} is greater than the number of 
linear mixed submodels, returns a list of numeric vectors for all outcomes.
}
\description{
Extracts estimated fixed effects for the longitudinal processes from a fitted joint model.
}
\examples{
\dontrun{
# linear mixed model fit
fit_lme1 <- lme\(log(serBilir) ~ year:sex + age,
                random = ~ year | id, data = pbc2)

fit_lme2 <- lme(prothrombin ~ sex, 
                random = ~ year | id, data = pbc2)

# cox model fit
fit_cox <- coxph(Surv(years, status2) ~ age, data = pbc2.id)

# joint model fit
fit_jm <- jm(fit_cox, list(fit_lme1, fit_lme2), time_var = "year")

# fixed effects for all linear mixed submodels 
fixef(fit_jm, outcome = 3)
}
}
\seealso{
\code{\link{coef}}, \code{\link{ranef}}, \code{\link{jm}}.
}
\author{
Dimitris Rizopoulos, \email{d.rizopoulos@erasmusmc.nl}.
}
