% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_methods.R
\name{gelman_diag.jm}
\alias{gelman_diag.jm}
\title{Gelman and Rubin's Convergence Diagnostic for Joint Models}
\usage{
\method{gelman_diag}{jm}(
  object,
  parm = c("all", "betas", "sigmas", "D", "bs_gammas", "tau_bs_gammas", "gammas",
    "alphas"),
  ...
)
}
\arguments{
\item{object}{an object inheriting from class \code{"jm"}.}

\item{parm}{either \code{"all"} or one specific joint model parameter of interest.}

\item{...}{further arguments passed to \code{\link[coda]{gelman.diag}}.}
}
\value{
A list of \code{gelman.diag} objects. An object of class 
\code{gelman.diag} is a list with the elements:
\tabular{ll}{
\code{psrf} \tab a list containing the point estimates of the potential 
scale reduction factor (labelled \code{Point est.}) and their upper 
confidence limits (labelled \code{Upper C.I.}). \cr
\code{mpsrf} \tab the point estimate of the multivariate potential scale 
reduction factor. This is NULL if the parameter is univariate.
}
}
\description{
Calculates the potential scale reduction factor for the estimated parameters
in a fitted joint model, together with the upper confidence limits.
}
\examples{
\dontrun{
# linear mixed model fit
fit_lme <- lme(sqrt(CD4) ~ obstime * drug, 
random = ~ 1 + obstime | patient, data = aids)

# cox model fit
fit_cox <- coxph(Surv(Time, death) ~ drug, data = aids.id)

# joint model fit
fit_jm <- jm(fit_cox, fit_lme, time_var = "obstime")

# joint model convergence diagnostic
gelman_diag(fit_jm, "all")
}
}
\seealso{
\code{\link[coda]{gelman.diag}}, \code{\link{jm}}.
}
\author{
Dimitris Rizopoulos, \email{d.rizopoulos@erasmusmc.nl}.
}
