% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_methods.R
\name{ggtraceplot.jm}
\alias{ggtraceplot.jm}
\title{Trace plot of MCMC output for Joint Models using ggplot2}
\usage{
\method{ggtraceplot}{jm}(
  object,
  parm = c("all", "betas", "sigmas", "D", "bs_gammas", "tau_bs_gammas", "gammas",
    "alphas"),
  size = 1,
  alpha = 0.8,
  theme = c("standard", "catalog", "metro", "pastel", "beach", "moonlight", "goo",
    "sunset"),
  grid = FALSE,
  gridrows = 3,
  gridcols = 1,
  ...
)
}
\arguments{
\item{object}{an object inheriting from class \code{"jm"}.}

\item{parm}{a character string with either the value \code{"all"} or one specific joint model group of parameters of interest. Defaults to \code{'all'}. \cr}

\item{size}{the width of the traceplot line in mm. Defaults to 1.}

\item{alpha}{the opacity level of the traceplot line. Defaults to 0.8.}

\item{theme}{a character string specifying the color theme to be used. Possible options are \code{'standard'}, \code{'catalog'}, \code{'metro'}, \code{'pastel'}, \code{'beach'}, \code{'moonlight'}, \code{'goo'}, \code{'sunset'}. \cr}

\item{grid}{logical; defaults to \code{FALSE}. If \code{TRUE} the plots are returned in grids split over multiple pages. For more details see the documentation for \code{\link[gridExtra:arrangeGrob]{gridExtra::marrangeGrob()}}. \cr}

\item{gridrows}{number of rows per page for the grid. Only relevant when using \code{grid = TRUE}. Defaults to 3.}

\item{gridcols}{number of columns per page for the grid. Only relevant when using \code{grid = TRUE}. Defaults to 1.}

\item{...}{additional arguments; currently none is used.}
}
\description{
Plots the evolution of the estimated parameter vs. iterations in a 
fitted joint model using ggplot2.
}
\examples{
\dontrun{
# linear mixed model fit
fit_lme <- lme(sqrt(CD4) ~ obstime * drug, random = ~ 1 + obstime | patient, 
               data = aids)

# cox model fit
fit_cox <- coxph(Surv(Time, death) ~ drug, data = aids.id)

# joint model fit
fit_jm <- jm(fit_cox, fit_lme, time_var = "obstime")

# trace plot for the fixed effects in the linear mixed submodel 
ggtraceplot(fit_jm, parm = "betas")
ggtraceplot(fit_jm, parm = "betas", grid = TRUE)
}
}
\seealso{
\code{\link{traceplot}}, 
\code{\link{jm}}.
}
\author{
Dimitris Rizopoulos, \email{d.rizopoulos@erasmusmc.nl}.
}
