% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_methods.R
\name{model.frame.jm}
\alias{model.frame.jm}
\title{Model.frame method for Joint Models}
\usage{
\method{model.frame}{jm}(
  object,
  process = c("longitudinal", "event"),
  type = c("fixed", "random"),
  ...
)
}
\arguments{
\item{object}{an object inheriting from class \code{"jm"}.}

\item{process}{which submodel to recreate the model frame, i.e.,
\tabular{ll}{
\code{longitudinal} \tab linear mixed model(s). \cr
\code{survival} \tab survival model.   
}}

\item{type}{which effects, fixed or random, to select, when 
\code{process = longitudinal}.}

\item{...}{additional arguments; currently none is used.}
}
\value{
If \code{process = "longitudinal"}, a list of the model frames used 
in the linear mixed model(s). If \code{process = "event"}, the model frame 
used in the survival model.
}
\description{
Creates the model frame from a fitted joint model.
}
\examples{
\dontrun{
# linear mixed model fit
fit_lme1 <- lme(log(serBilir) ~ year:sex + age, random = ~ year | id, 
                data = pbc2)

fit_lme2 <- lme(prothrombin ~ sex, random = ~ year | id, data = pbc2)

# cox model fit
fit_cox <- coxph(Surv(years, status2) ~ age, data = pbc2.id)

# joint model fit
fit_jm <- jm(fit_cox, list(fit_lme1, fit_lme2), time_var = "year")

# model frame for the fixed effects terms in the linear mixed submodels
model.frame(fit_jm, process = "longitudinal", type = "fixed")
}
}
\seealso{
\code{\link[stats]{model.frame}}, \code{\link{jm}}.
}
\author{
Dimitris Rizopoulos, \email{d.rizopoulos@erasmusmc.nl}.
}
