% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_methods.R
\name{model.matrix.jm}
\alias{model.matrix.jm}
\title{Design Matrices for Linear Mixed Submodels in Joint Models}
\usage{
\method{model.matrix}{jm}(object, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{"jm"}.}

\item{...}{additional arguments; currently none is used.}
}
\value{
A list of the design matrices for the linear mixed submodels.
}
\description{
Creates the design matrices for linear mixed submodels from a fitted joint 
model.
}
\examples{
\dontrun{
# linear mixed model fit
fit_lme1 <- lme(log(serBilir) ~ year:sex + age, random = ~ year | id, 
                data = pbc2)

fit_lme2 <- lme(prothrombin ~ sex, random = ~ year | id, 
                data = pbc2)

# cox model fit
fit_cox <- coxph(Surv(years, status2) ~ age, data = pbc2.id)

# joint model fit
fit_jm <- jm(fit_cox, list(fit_lme1, fit_lme2), time_var = "year")

# linear mixed models design matrices
model.matrix(fit_jm)
}
}
\seealso{
\code{\link[stats]{model.matrix}}, \code{\link{jm}}.
}
\author{
Dimitris Rizopoulos, \email{d.rizopoulos@erasmusmc.nl}.
}
