% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_methods.R
\name{ranef.jm}
\alias{ranef.jm}
\title{Random Effects Estimates for Linear Mixed Submodels in Joint Models}
\usage{
\method{ranef}{jm}(object, outcome = Inf, post_vars = FALSE, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{"jm"}.}

\item{outcome}{the linear mixed submodel to extract the estimated fixed effects.
If greater than the total number of linear mixed submodels, extracts from all 
of them.}

\item{...}{additional arguments; currently none is used.}
}
\value{
A numeric matrix with rows denoting the individuals and columns the random 
effects. If \code{postVar = TRUE}, the numeric matrix has an extra attribute 
“postVar".
}
\description{
Extracts estimated random effects for the longitudinal processes from a fitted joint model.
}
\examples{
\dontrun{
# linear mixed model fits
fit_lme1 <- lme(log(serBilir) ~ year:sex + age, random = ~ year | id, 
                data = pbc2)

fit_lme2 <- lme(prothrombin ~ sex, random = ~ year | id, data = pbc2)

# cox model fit
fit_cox <- coxph(Surv(years, status2) ~ age, data = pbc2.id)

# joint model fit
fit_jm <- jm(fit_cox, list(fit_lme1, fit_lme2), time_var = "year")

# random effects from all linear mixed submodels 
ranef(fit_jm)
}
}
\seealso{
\code{\link{coef}}, \code{\link{fixef}}, \code{\link{jm}}.
}
\author{
Dimitris Rizopoulos, \email{d.rizopoulos@erasmusmc.nl}.
}
