% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_methods.R
\name{terms.jm}
\alias{terms.jm}
\title{Joint Model Terms}
\usage{
\method{terms}{jm}(
  object,
  process = c("longitudinal", "event"),
  type = c("fixed", "random"),
  ...
)
}
\arguments{
\item{object}{an object inheriting from class \code{"jm"}.}

\item{process}{which submodel to extract the terms, i.e.,
\tabular{ll}{
\code{longitudinal} \tab linear mixed model(s). \cr
\code{survival} \tab survival model.   
}}

\item{type}{which effects, fixed or random, to select, when 
\code{process = "longitudinal"}.}

\item{...}{additional arguments; currently none is used.}
}
\value{
If \code{process = "longitudinal"}, a list of the terms object(s) for
 the linear mixed model(s). If \code{process = "event"}, the terms object 
 for the survival model.
}
\description{
Extracts the terms objects from a fitted joint model.
}
\examples{
\dontrun{
# linear mixed model fit
fit_lme <- lme(log(serBilir) ~ year * sex, random = ~ year | id, data = pbc2)

# cox model fit
fit_cox <- coxph(Surv(years, status2) ~ age, data = pbc2.id)

# joint model fit
fit_jm <- jm(fit_cox, fit_lme, time_var = "year")

# fixed effects terms in the linear mixed model
terms(fit_jm, process = "longitudinal", type = "random")
}
}
\seealso{
\code{\link{model.frame}}, \code{\link{jm}}.
}
\author{
Dimitris Rizopoulos, \email{d.rizopoulos@erasmusmc.nl}.
}
