% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_methods.R
\name{traceplot.jm}
\alias{traceplot.jm}
\title{Trace plot of MCMC output for Joint Models}
\usage{
\method{traceplot}{jm}(
  object,
  parm = c("all", "betas", "sigmas", "D", "bs_gammas", "tau_bs_gammas", "gammas",
    "alphas"),
  ...
)
}
\arguments{
\item{object}{an object inheriting from class \code{"jm"}.}

\item{parm}{either \code{"all"} or one specific joint model parameter of interest.}

\item{...}{further arguments passed to \code{\link[coda]{traceplot}}.}
}
\description{
Plots the evolution of the estimated parameter vs. iterations in a 
fitted joint model.
}
\examples{
\dontrun{
# linear mixed model fit
fit_lme <- lme(sqrt(CD4) ~ obstime * drug, random = ~ 1 + obstime | patient, 
               data = aids)

# cox model fit
fit_cox <- coxph(Surv(Time, death) ~ drug, data = aids.id)

# joint model fit
fit_jm <- jm(fit_cox, fit_lme, time_var = "obstime")

# trace plot for the fixed effects in the linear mixed submodel 
traceplot(fit_jm, parm = "betas")
}
}
\seealso{
\code{\link[coda]{traceplot}}, \code{\link{ggtraceplot}}, 
\code{\link{jm}}.
}
\author{
Dimitris Rizopoulos, \email{d.rizopoulos@erasmusmc.nl}.
}
